+++ src/components/Crowdsale/CrowdsaleList/CrowdsaleList.tsx
@@ -1,21 +1,31 @@
 import React from 'react';
-import { Flex, Text } from 'rebass';
+import { Flex, Text, Image } from 'rebass';
 import '../../../assets/styles/NewToken.css';
 import CrowdsaleListItem from './CrowdsaleListItem';
 import { useTranslation } from 'react-i18next';
 import CrowdsaleListPlaceholder from './CrowdsaleListPlaceholder';
 import { ICrowdsaleData } from 'src/interfaces';
 import { CrowdsaleSortEnum } from 'src/api/middleware';
+import { useSelector } from 'react-redux';
+import EmptyImg from '../../../images/empty.png';
 
 interface IProps {
   crowdsaleList: ICrowdsaleData[];
-  fetchMore: any;
+  fetchMore?: any;
   tokenList: any;
   limit: number;
 }
 
 const CrowdsaleList: React.FC<IProps> = ({ crowdsaleList, tokenList, fetchMore, limit }) => {
   const { t } = useTranslation();
+
+  const { txnLoading } = useSelector(
+    ({ chain }: any) => {
+      return {
+        txnLoading: chain.txnLoading,
+      };
+    },
+  );
 
   const handleScroll = ({ currentTarget }: any, fetchMore: any) => {
     if (
@@ -27,7 +37,7 @@ const CrowdsaleList: React.FC<IProps> = ({ crowdsaleList, tokenList, fetchMore, 
           skip: 0,
           filter: {},
           sort: CrowdsaleSortEnum.DESC,
-          limit: limit + 20,
+          limit: limit + 10,
         },
       });
     }
@@ -46,13 +56,19 @@ const CrowdsaleList: React.FC<IProps> = ({ crowdsaleList, tokenList, fetchMore, 
         {t('marketplace.label')}
       </Text>
       <div style={{ overflow: 'scroll', height: '85vh' }}>
-        <Flex flexDirection="column">
-          {crowdsaleList && crowdsaleList.length > 0 ? (
-            crowdsaleList.map((crowdsale: ICrowdsaleData, index: number) => (
+        <Flex flexDirection="column" height="100%">
+          {txnLoading && crowdsaleList.length === 0 ? (
+            <CrowdsaleListPlaceholder counter={2} />
+          ) : !txnLoading && crowdsaleList.length === 0 ? (
+            <Flex paddingBottom="125px" width="212px" margin="auto" flexDirection="column">
+              <Image marginBottom="35px" src={EmptyImg} />
+              <Text width="195px" textAlign="center">{t('marketplace.no_crowdsale')}</Text>
+            </Flex>
+
+          ) : (
+            crowdsaleList && crowdsaleList.length > 0 && crowdsaleList.map((crowdsale: ICrowdsaleData, index: number) => (
               <CrowdsaleListItem key={index} crowdsale={crowdsale} tokenList={tokenList} />
             ))
-          ) : (
-            <CrowdsaleListPlaceholder counter={2} />
           )}
         </Flex>
       </div>